flashを表示させる
