+++ app/helpers/my_page/trade_logs_helper.rb
@@ -1,5 +1,5 @@
 module MyPage::TradeLogsHelper
   def change_sender_name_from_admin(trade)
-    trade.sender.type == "AdminUser" ? "MPay" : trade.sender.name
+    trade.sender.AdminUser? ? "MPay" : trade.sender.name
   end
 end
