+++ app/javascript/clipboard/controllers/clipboard_controller.js
