+++ app/javascript/flash/controllers/flash_controller.js
