+++ app/javascript/my_page/controllers/friend_search_controller.js
