+++ app/javascript/packs/application.js
@@ -1,20 +1,13 @@
-/* eslint no-console:0 */
-// This file is automatically compiled by Webpack, along with any other files
-// present in this directory. You're encouraged to place your actual application logic in
-// a relevant structure within app/javascript and only use these pack files to reference
-// that code so it'll be compiled.
-//
-// To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
-// layout file, like app/views/layouts/application.html.erb
+import Rails from 'rails-ujs'
+import { Application } from 'stimulus'
+import ClipboardController from '../clipboard/controllers/clipboard_controller'
+import FlashController from '../flash/controllers/flash_controller'
+import FriendSearchController from '../my_page/controllers/friend_search_controller'
+import RelationshipsController from '../relationships/controllers/relationships_controller'
 
-
-// Uncomment to copy all static images under ../images to the output folder and reference
-// them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
-// or the `imagePath` JavaScript helper below.
-//
-// const images = require.context('../images', true)
-// const imagePath = (name) => images(name, true)
-
-console.log('Hello World from Webpacker')
-
-import "controllers"
+const application = Application.start()
+application.register('clipboard', ClipboardController)
+application.register('flash', FlashController)
+application.register('friend-search', FriendSearchController)
+application.register('relationships', RelationshipsController)
+Rails.start()
