+++ app/views/layouts/application.html.erb
@@ -6,7 +6,6 @@
     <%= csp_meta_tag %>
 
     <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
     <%= javascript_pack_tag "application" %>
     <link href='https://fonts.googleapis.com/icon?family=Material+Icons' rel='stylesheet'>
   </head>
