+++ package.json
@@ -3,6 +3,8 @@
   "private": true,
   "dependencies": {
     "@rails/webpacker": "^4.0.7",
+    "global": "^4.4.0",
+    "rails-ujs": "^5.2.3",
     "stimulus": "^1.1.1"
   },
   "devDependencies": {
